LPS-108343 Rename files: remove .es
