+++ modules/dxp/apps/analytics/analytics-reports-web/src/main/resources/META-INF/resources/js/utils/numberFormat.js
